+++ Tests/Functional/Command/FlushQueueCommandTest.php
@@ -83,12 +83,12 @@ class FlushQueueCommandTest extends AbstractCommandTests
             'Flush Finished' => [
                 'mode' => 'finished',
                 'expectedOutput' => 'All entries in Crawler queue, with status: "finished" will be flushed',
-                'expectedCount' => 7,
+                'expectedCount' => 8,
             ],
             'Unknown mode' => [
                 'mode' => 'unknown',
                 'expectedOutput' => 'No matching parameters found.',
-                'expectedCount' => 14,
+                'expectedCount' => 15,
             ],
         ];
     }