+++ Tests/Unit/Hooks/DataHandlerHookTest.php
@@ -4,6 +4,21 @@ declare(strict_types=1);
 
 namespace AOE\Crawler\Tests\Unit\Hooks;
 
+/*
+ * (c) 2020 AOE GmbH <dev@aoe.com>
+ *
+ * This file is part of the TYPO3 Crawler Extension.
+ *
+ * It is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License, either version 2
+ * of the License, or any later version.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ *
+ * The TYPO3 project - inspiring people to share!
+ */
+
 use AOE\Crawler\Api\CrawlerApi;
 use AOE\Crawler\Domain\Repository\QueueRepository;
 use AOE\Crawler\Hooks\DataHandlerHook;
@@ -53,7 +68,7 @@ class DataHandlerHookTest extends UnitTestCase
 
         $dataHandlerHook->addFlushedPagesToCrawlerQueue(
             [
-                'pageIdArray' => [1, 2]
+                'pageIdArray' => [0, 1, 2],
             ],
             $dataHandler
         );